+++ ignite/pkg/cosmosgen/cosmosgen.go
@@ -17,9 +17,8 @@ type generateOptions struct {
 	includeDirs []string
 	gomodPath   string
 
-	jsOut               func(module.Module) string
-	jsIncludeThirdParty bool
-	tsClientRootPath    string
+	jsOut            func(module.Module) string
+	tsClientRootPath string
 
 	vuexOut      func(module.Module) string
 	vuexRootPath string
@@ -50,26 +49,23 @@ func WithTSClientGeneration(out ModulePathFunc, tsClientRootPath string) Option 
 	}
 }
 
-func WithVuexGeneration(includeThirdPartyModules bool, out ModulePathFunc, vuexRootPath string) Option {
+func WithVuexGeneration(out ModulePathFunc, vuexRootPath string) Option {
 	return func(o *generateOptions) {
 		o.vuexOut = out
-		o.jsIncludeThirdParty = includeThirdPartyModules
 		o.vuexRootPath = vuexRootPath
 	}
 }
 
-func WithComposablesGeneration(includeThirdPartyModules bool, out ModulePathFunc, composablesRootPath string) Option {
+func WithComposablesGeneration(out ModulePathFunc, composablesRootPath string) Option {
 	return func(o *generateOptions) {
 		o.composablesOut = out
-		o.jsIncludeThirdParty = includeThirdPartyModules
 		o.composablesRootPath = composablesRootPath
 	}
 }
 
-func WithHooksGeneration(includeThirdPartyModules bool, out ModulePathFunc, hooksRootPath string) Option {
+func WithHooksGeneration(out ModulePathFunc, hooksRootPath string) Option {
 	return func(o *generateOptions) {
 		o.hooksOut = out
-		o.jsIncludeThirdParty = includeThirdPartyModules
 		o.hooksRootPath = hooksRootPath
 	}
 }
@@ -148,12 +144,20 @@ func Generate(ctx context.Context, cacheStorage cache.Storage, appPath, protoDir
 			return err
 		}
 
-		// Update Vue app dependeciens when Vuex stores are generated.
+		// Update Vuex store dependecies when Vuex stores are generated.
+		// This update is required to link the "ts-client" folder so the
+		// package is available during development before publishing it.
+		if err := g.updateVuexDependencies(); err != nil {
+			return err
+		}
+
+		// Update Vue app dependecies when Vuex stores are generated.
 		// This update is required to link the "ts-client" folder so the
 		// package is available during development before publishing it.
 		if err := g.updateVueDependencies(); err != nil {
 			return err
 		}
+
 	}
 
 	if g.o.composablesRootPath != "" {
@@ -161,7 +165,7 @@ func Generate(ctx context.Context, cacheStorage cache.Storage, appPath, protoDir
 			return err
 		}
 
-		// Update Vue app dependencies when Vuex stores are generated.
+		// Update Vue app dependencies when Vue composables are generated.
 		// This update is required to link the "ts-client" folder so the
 		// package is available during development before publishing it.
 		if err := g.updateComposableDependencies("vue"); err != nil {
@@ -173,7 +177,7 @@ func Generate(ctx context.Context, cacheStorage cache.Storage, appPath, protoDir
 			return err
 		}
 
-		// Update React app dependencies when Vuex stores are generated.
+		// Update React app dependencies when React hooks are generated.
 		// This update is required to link the "ts-client" folder so the
 		// package is available during development before publishing it.
 		if err := g.updateComposableDependencies("react"); err != nil {