+++ ignite/templates/module/create/ibc.go
@@ -20,7 +20,7 @@ import (
 func NewIBC(replacer placeholder.Replacer, opts *CreateOptions) (*genny.Generator, error) {
 	var (
 		g        = genny.New()
-		template = xgenny.NewEmbedWalker(fsIBC, "ibc/", opts.AppPath)
+		template = xgenny.NewEmbedWalker(fsIBC, "files/ibc/", opts.AppPath)
 	)
 
 	g.RunFn(genesisModify(replacer, opts))